cast to int for padding

+++ demos/sky_model.py
@@ -17,7 +17,7 @@ class ErositaSky:
         grid_info = self.config['grid']
         tel_info = self.config['telescope']
         self.position_space = ift.RGSpace(2*(self.config['grid']['npix'],), distances=[2.0 * tel_info['fov'] / grid_info['npix']])
-        extended_size = self.config['grid']['padding_ratio'] * self.position_space.shape[0]
+        extended_size = int(self.config['grid']['padding_ratio'] * self.position_space.shape[0])
         self.extended_space = ift.RGSpace(2*(extended_size,), distances=self.position_space.distances)
         self.pad = ift.FieldZeroPadder(self.position_space, self.extended_space.shape)
